requirements.txt
beautifulsoup4==4.9.3
Flask==2.0.1
httpx==0.18.2
pytest-flask==1.2.0
